Initial Commit of Symbol-Type Graph (STG)
The STG (symbol-type graph) is an ABI representation and this project
contains tools for the creation and comparison of such representations.
At present parsers exist for libabigail's ABI XML (C types only) and
BTF. The ABI diff tool, stgdiff, supports multiple reporting options.
This software currently depends on libabigail for ELF symbol
functionality, on libxml2 for XML parsing and on Linux UAPI headers for
BTF types.

Co-authored-by: Maria Teguiani <teguiani@google.com>
Signed-off-by: Maria Teguiani <teguiani@google.com>
Signed-off-by: Giuliano Procida <gprocida@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>