stgdiff: tighten criteria for being a diff-holding node
For the flat and small format reports, the concept of "diff-holding"
node governs which nodes can appear at the top-level of a report.

We want these to be entities that can be considered to have changed
intrinisically. In particular, for user-defined types, we should
consider entities to be distinct if the names differ.

This change does this for structs, union and enums. Additionally, a
switch between being a struct and a union now prevents a diff node
from being given top-level status.

All this precludes any concept of renaming which to handle properly
would require heuristics to determine whether the before / after
types were sufficiently similar to warrant different reporting.

The other kind of user-defined types, typedefs, will be tweaked in a
follow-up change.

PiperOrigin-RevId: 407589851
Change-Id: I759bef809b0a68b22ea5f1507802580bed274b19