Adopt generic system include paths for libabigail headers
When including libabigail headers, the idiomatic way is to use the
prefixed system include, usually residing in /usr/include/libabigail and
included as -I/usr/include/libabigail:
  #include <abg-*>

For the soong build, we can set up this include path easily by exporting
the required headers from libabigail. There is one caveat: headers
included via libabigail/src/abg-* are usually not part of the
distribution and unlikely to be found in /usr/include/libabigail or
equivalent. This is somewhat a special case where we are using an
internal header (and it is actually perfectly fine).
For soong we can set up the build as if it were exposed by the
libabigail library and we should be good then.

For now, teach copybara to rewrite headers accordingly.

PiperOrigin-RevId: 410224217
Change-Id: I5f8e64ffc9f8bf3bdbb6d1755d0fff6082022690