BTF reader: validate all BTF subsection bounds
Arithmetic on fixed width quantities (uint32_t) can lead to
wrap-around and nonsensical bounds.

This change ensures all subsections are in the right order and are
non-overlapping with respect to each other and the BTF section limits

This change also ensures the type and string section bounds are
calculated exactly once.

The code is a bit tidier as well with the MemoryRange helper used more
extensively.

#fuzzit

PiperOrigin-RevId: 410248352
Change-Id: Ie47462b7675895f290a189b31ac010c59d7748c1