Improve typedef diff attribution
The flat and small format reports need diffs to be correctly
attributed to user-defined entities in order to produce meaningful
diff output.

In the simple case of a typedef whose definition has changed,
`stgdiff` correctly attributes the difference to the typedef.

However, should a typedef change to a non-typedef or vice versa, the
diff would be attributed to this comparison, rather than the ancestral
user-defined type comparison.

With this commit, diffs are attributed to a comparison node involving
typedefs **only if** both sides are typedefs of the same name. Typedef
chains are still fully resolved before attempting further comparisons.

PiperOrigin-RevId: 426384877
Change-Id: I05ea2671f8374fd73dffe4b57404b4f88ba294cf