Refer to root nodes as root nodes
At present, a STG graph root node always refers to the symbol table of
an ABI and the only code that is sensitive to this choice is the
unrolling of top-level diffs in the reporting code which is still
agnostic to the node kinds being compared.

Changing mentions from "symbols" to "root" makes this generality more
obvious.

PiperOrigin-RevId: 435316048
Change-Id: Ia9d745a5497703d5dfc957dc196d87fd02d3ab27