Move graph mutation functionality into Graph base class
The graph building code in each reader used `Graph` as an abstract base
class. This had the convenience of avoiding lifetime issues but didn't
really offer extra flexibility as a node array reference was still
handed to every node constructor. All functionality to add nodes to the
array etc. was duplicated in each reader.

This change is a first step towards making `Graph` into a concrete
type and so allow uniform manipulation of graphs. Almost all of the
graph building functionality (including the node array) is now in
`Graph` itself, but that is still being used as a base class and the
node array is still directly exposed.

PiperOrigin-RevId: 435316530
Change-Id: Ifda20fc680841288af14ecf48987cd5dd06ee92b