Make state the first argument of comparison functions
The `State` type is the context holder for comparison operations and
will soon include a `Graph` reference. The functions are:

* `Removed`
* `Added`
* `Compare`
* `Equals`

The argument lists all now look like:

1. `state`
2. nodes

This should make for more readable and maintainable code.

A related change will follow to consolidate and pass reporting context
in a single structure.

PiperOrigin-RevId: 435316578
Change-Id: Iad5eccf73280a6a7f297ccca583fba9688537ee8