add node construction helper and make node array private
This commit introduces the template function `Make` which forwards its
arguments to `std::make_unique`. It also passes `Graph`'s own `types_`
array, though this will be dropped in a later change.

This change eliminates a lot of boilerplate around node creation. It
removes most references to `types_` and makes it private to `Graph`.

PiperOrigin-RevId: 436765769
Change-Id: Ib0b0caaf607f0b1eb41d9f103abec83db27c97a5