Split definitions of recursive functions on Type
The virtual functions:

* `ResolveQualifiers`
* `ResolveTypedef`

were recursively-defined member functions of `Type`. This commit
replaces them with pairs of functions:

* `ResolveQualifiers` / `ResolveQualifier`
* `ResolveTypedefs` / `ResolveTypedef`

where the first of each pair is a free wrapper function and the second
is a virtual worker function.

This scheme makes it possible to separate node lookup from the real
work. Node lookup will no longer be an integral part of the `Type`
base class. More significantly `Type` methods will have no way of
returning "self" as a node `Id` as `Type` has no awareness of this
identity, but the wrapper function can hold onto this.

PiperOrigin-RevId: 437251584
Change-Id: Ic9ca733f7036027f3f86cefe1d558499b62ccb8a