drop parameter names from function type descriptions
Parameter names are stored in STG's internal representation but they
are not part of the type system and do not feature in type
comparisons. They are, however, used when generating descriptions of
function types and in diff reports involving function parameters.

The use of parameter names in generated type descriptions is
problematic for various reasons:

* The types are needlessly verbose.
* Equal types may have multiple differing descriptions.
* The redundant name information is, in general, unstable.

This sometimes also pollutes STG diff reports when input XML files
flip between different function declarations, some of which may have
names that differ or are absent altogether.

For exported functions with CRC-only changes, parameter name changes
also upset post-processing.

PiperOrigin-RevId: 438048427
Change-Id: I7a05972a1f0acdf429918b25d54b71cc2e285457