adding Dockerfile, Makefile, and documentation
There is currently no instruction or build file to make it easy
to develop STG tools. This changeset adds a basic Makefile, automated
container build, and associated documentation as the start of work
to help with that. Note that it is not intended to be a final
solution, but to empower interested developers to build STG tools.
Note that the Dockerfile uses a Fedora base, and this is done for
the libabigail build, which would take quite a bit of extra work
to figure out on a debian slim base. This Dockerfile base can be
updated to debian-slim (a more stable container) when the
libabigail dependency is removed.

Signed-off-by: Vanessa Sochat <vsochat@gmail.com>
Change-Id: Ifa6528cc738e7a694a3b4470894ab9d3484d8394