Add workaround for `is-declaration-only` classes.
DWARF's `AT_DW_declaration` means "incomplete" type.

Compiler writers have (ab)used this to emit partial information for certain base classes, in order to save time and space, at the obvious expense of incomplete information, including loss of type size.

Libabgail tries to preserve the original intent of the attribute and calls it `is-declaration-only` in XML. The name is misleading for structure types and the Abigail reader needs to be more careful about how it handles them.

PiperOrigin-RevId: 456724939
Change-Id: I4d2e251659abf5d8da5a23d7448b04ffad4593e9