Match symbols by object path and name
In the case of corpus groups, we should consider symbols in different
modules to be independent (and not duplicates if they happen to have
the same name).

For the moment, we won't model corpus groups directly, but we can make
sure to scope symbols name by corpus path and mention the path in
symbol diffs (if it is not "vmlinux").

PiperOrigin-RevId: 459449265
Change-Id: I1a5fd1571c8bfb4262e7c23d14b2c7c708562040