ELF loader: replace `GetSymbolName` with more general `GetString`
The pattern "go to the strings section and get me string with some id"
is not unique to the symbol table.

This change introduces method `GetString`, which is basically a
wrapper for `elf_strptr`. It returns a `std::string_view` referring to
ELF file memory, so there will be no copying and constant memory
consumption when storing ELF strings.

PiperOrigin-RevId: 466325363
Change-Id: I78ee263f1dd9013d1dae44cd86c68594cf4e4fe5