ELF loader: factor "number of entries" calculation into own function
This change introduces `GetNumberOfEntries` method which converts
`sh_entsize` and `sh_size` into the number of entries for sections
that hold a table of fixed-size entries.

This change also implements printing of some ELF section types to
provide context in error messages. The method only supports the
section types used in `ElfLoader`.

PiperOrigin-RevId: 466325606
Change-Id: Ia17fdc72b6e1ca42a757c5d57fd5d7b606ad063b