Improve descriptions involving pointer and reference types
The existing pretty printer was fairly conservative when emitting
unbracketed type expressions. It always inserted a space for "left"
syntax (i.e., not `[...]` or `(...)`) that wasn't immediately preceded
by a `(`.

However, in practice such syntax is either a pointer or reference sign
or an atomic word token. It's only word tokens that need to be
protected from unwanted joining (as references to references are
fortunately illegal).

|old                           |new                         |
|------------------------------|----------------------------|
|`int * * x`                   |`int** x`                   |
|`volatile int * const * const`|`volatile int* const* const`|
|`const struct foo * * j`      |`const struct foo** j`      |

Qualifiers already have special treatment and are not affected. It's
possible more could be done with them.

PiperOrigin-RevId: 478535181
Change-Id: I6763f22e205d225547f8c72e7627261bb0888e30