SCC finder: allow non default-constructible node types
The use of `std::vector::resize` to truncate the "open" nodes vector
imposed an unnecessary restriction on the node type as `resize` can
also be used to grow vectors.

This change just uses a direct call to `erase` instead.

A couple of other small improvements are included here:

* The `pop_back` of the root index vector is moved to immediately
after the last element test, mainly for readability.
* The `scc` vector has the exact space required reserved.

PiperOrigin-RevId: 479308115
Change-Id: I24b7a90c95f79f08dfe48ecf29ee7ddd2ee01b63