Improve safety of operator<< for Integer::Encoding
Having a static array of integer encoding descriptions made for short
code, but risked the `enum` and array of descriptions getting out of
sync without any compiler warnings.

This change makes the output operator for `Integer::Encoding` use a
`switch` statement so that any missing cases result in warnings; it
also eliminates a type cast.

PiperOrigin-RevId: 479544387
Change-Id: I6863e33896d3a81a93c36ea21d1d2893badacdf9