Abigail reader: work around missing mangled method names
We sometimes see methods (usually, or always, constructors and
destructors) lacking a `mangled-name` attribute in ABI XML.

The corresponding DWARF `linkage_name` attribute should always be
present in complete method DIEs. However, sometimes the information is
split between a declaration DIE (missing a linkage name) and another
specification DIE referring to the declaration DIE (containing a
linkage name), and sometimes the latter is not picked up by
libabigail.

This change replaces the missing information with `{missing}` as
there's not much else we can do.

PiperOrigin-RevId: 480680123
Change-Id: I26a41ba6cb3e40e4c3fb4cf48120a32db76d77da