reporting: plain report: move "being/already reported" to separate line
Earlier, "being/already reported" was being added as a suffix to a
comparison which was already being reported or had been reported. Now,
it won't be a suffix and rather be printed on the next line (with extra
indentation to denote that this information is regarding the comparison
right before it). Thus, this change helps simplify and reduce the
amount of information being printed on a single line.

Note, that there is no change to the reporting functionality and all
information which was being printed is the same. This is rather a part
of internal cleanup of reporting.

PiperOrigin-RevId: 487784486
Change-Id: Id8b7a4dd4bfd30996eca5ebbfbe512a7fd8c9c3a