reporting: print edge diffs with their corresponding comparisons
Move away from printing edge diffs and then calling
`PlainPrint`/`FlatPrint` for the next comparison to passing the edge
diff as a prefix to the next `PlainPrint`/`FlatPrint` call. This helps
making sure that all calls to print comparisons own the entire task of
printing a diff line in the report. Hence it makes the printing logic
simpler and more contained within `PrintComparison`.

PiperOrigin-RevId: 487784596
Change-Id: Ib8957eaecf906be913a342bb6f8cff58fc5a26e9