reporting: print newlines following diffs in `PrintComparison`
Ensure that newlines are handled by `PrintComparison` itself and any
other logic to print lines in reports need not worry about the handling
of newlines in different parts of the code. In other words, all printing
code will handle its own newlines moving forward, and not tangle with
other lines in a diff report.

PiperOrigin-RevId: 487784695
Change-Id: Id51fdbe69e641588d20859652fdbe0922cd06bf4