Replace `GetEnumNames` method with a standalone function
This method was used to generate input for `PairUp` and was the last
remaining method in any of the node types, except for the base class.

It was only ever called on `Enum` values corresponding to fully-defined
enumerations, so its replacement is just called with the enumerators
themselves.

The replacement is called `MatchingKeys` as there is another overload
that does a similar thing with different input.

PiperOrigin-RevId: 488646972
Change-Id: Icc9a3f6fd348bc00d6f573796d25d57a0aa3e05e