DWARF reading: support emitting nodes to STG
Add a method `AddProcessedNode` to allocate a fixed `Id` for
`dwarf::Entry`. It allows us to reference it in STG without the need to
process it. But to make sure that the processor for this entry will use
exactly that allocated `Id`, add a method `AddProcessedNode`, which
uses `GetIdForEntry` to populate the node with the preallocated `Id`.

Also it is useful to record the allocation of `Id` in the processed
result (`Types`) for postprocessing. For example, to check, that all ids
have an associated node.

PiperOrigin-RevId: 489981914
Change-Id: I6d8eeadf9ec63e184d4db304e9b6a0b3babe8125