make graph nodes mutable
This change allows all fields, including node ids, within non-`const` `Graph`
nodes to be updated.

The current graph representation supports node creation, but not node deletion
or modification.

Node deletion, in the absence of bugs, is purely a space optimisation as
deleted nodes should not be reachable anyway. We can implement it later if
necessary.

Without mutability, nodes can have drop-in node replacements which is one way of
"updating" node fields, but it's not enough to efficiently support applications
such as node/path deduplication which require mass edge repointing.

No graph exists in isolation. There are always going to be one or more
references outside the graph to nodes of interest. Any repointing of edges has
to apply to such external references as well. This means we cannot hide the
mappings altogether.

Without radical changes, there are a few general approaches to allowing edge
repointing:

* introduce a special `Redirect` node kind; this would require every single
  graph algorithm to act specially and locally whenever encountering it
* maintain repointing information wholly inside the graph data structure so that
  "downstream" algorithms can be oblivious - except to the extent that they care
  about node identity (i.e., any time node ids are compared or used as map keys)
* maintain repointing information wholly outside the graph data structure and
  require all "downstream" algorithms to be aware of this state so that they
  visit the right nodes
* give up node immutability and rewrite graph nodes by applying node mappings

More radical options include:

* split the concept of node identity into node references (like file
  descriptors) and node identities (like actual files) and maintain a mapping
  from reference to identity within the graph, exposed to the users that care
  about identity
* reify the concept of edge - almost the same as the preceding option, except
  that a reference is associated with a pair of nodes instead of just one

To preserve the simplicity both of concepts and of non-mutating algorithms (like
comparison and serialisation), we've decided to give up immutability and mutate
nodes locally where needed. This also avoids issues with associated edge
mappings changing over time but introduces the similar risk of mutating a node
that some algorithm may be traversing.

PiperOrigin-RevId: 491844883
Change-Id: I4055cf1fd2bd7333a6963fcbdd6bcb49bf1b2d08