equality: add node equality algorithm
The function object `Equals` determines whether two nodes are "equal" as values,
recursively. This means they are indistinguishable, in terms of having the same
node attributes and edge labels and having the child nodes reachable following
the same edges be indistinguishable etc.

The algorithm can be used with different equality caches to suit different
purposes. We are not sure what will be needed right now in terms of
functionality and performance and this will make it possible to try various
approaches and cross-check behaviour.

PiperOrigin-RevId: 491855214
Change-Id: I681692ff8d9ed2f653aaf5d21989f208823330b4