equality: make node equality algorithm header-only
This is parameterised by the equality cache implementation. There's not much
point implementing a bunch of canned cases. Each user will know what it wants
best. This means we need to expose the code in the header.

This change also makes `Equals` hold the equality cache as a reference. This is
in case its lifetime must exceed that of a single `Equals` instance.

PiperOrigin-RevId: 491918708
Change-Id: I832295f8b4fcf0285a1ad0014f6426daafba4a8a