comparison: generate member matching keys more carefully
When comparing

```
struct A {
  ...
  int x;
  int y;
  ...
};
```

with

```
struct A {
  ...
  struct {
    int x;
    int y;
    ...
  };
  ...
};
```

STG would try to match the `int x` member against the anonymous member because
it computes both to have a "matching key" of `x`, resulting in a rather confused
message about a member change. The matching key feature was designed to improve
diff reporting for types containing multiple anonymous members.

This change causes the matching keys for nested members to differ, breaking the
match. Note that there is no attempt to check whether the anonymous member is
ABI equivalent to replacing it with its own nested members.

PiperOrigin-RevId: 493267253
Change-Id: Ia7bd76b808186f585cd499e860cdb2d49c86e19c