DWARF reading: add a method to get abstract and string attributes
This method uses the `GetAttribute` function to get an "abstract"
`Dwarf_Attribute`. If the DWARF DIE has `DW_AT_specification` or the
`DW_AT_abstract_origin` attribute, the complete definition of this entry
may be in some other DIE. In this case the DIE is interpreted as if it
has "integrated" all attributes inside itself, which provides the most
complete information for the entry. That's why we have to use
`dwarf_attr_integrate`, which does exactly what is described above.

An abstract attribute is passed to the `dwarf_formstring`, which
extracts the string. If the attribute contains some other form, the
`dwarf_formstring` returns `nullptr`, which we interpret as error.

PiperOrigin-RevId: 493408584
Change-Id: I8e6fbd32276a56b3a9b818172552784aa44e8ab3