DWARF reading: process primitive types
Add support for DWARF entries tagged as `DW_TAG_base_type`. All
primitive types are required to have the name in attribute `DW_AT_name`.
A lot of other DWARF types have `DW_AT_name`, so move this code into a
separate function.

Also, primitive types should have a byte or a bit size and an encoding
and we should fail if the encoding has an unexpected value.

PiperOrigin-RevId: 493408968
Change-Id: Ia4d27dee57ca06c9776f08b3677c45797c03d3fe