fingerprint: good, fast node hashing algorithm
Testing shows this is good enough for node deduplication of graphs generated by
the Abigail reader. It may find other uses or be further refined.

On encountering a non-trivial SCC, it has to discard the hashing information
collected and assign the same fingerprint (the size of the SCC) to all nodes in
the SCC. This hasn't yet happened in practice, but the code is instrumented to
detect this.

PiperOrigin-RevId: 493834458
Change-Id: I86ec71861cca634768db68f0f1eb1b63df883e33