Use DWFL to support relocations in DWARF
DWARF5 adds an indirection layer `.debug_str_offsets` for string
attributes, but this section uses relocations. There is no dedicated
function in `elfutils` to apply relocations to ELF sections, so the only
choice left is to use `libdwfl`.

It handles the whole process of getting DWARF information from file or
memory:
* Opens and handles `Elf*`.
* Applies relocations to its sections.
* Opens and handles `Dwarf*`.

There is no impact on performance for reading DWARF entries, because
`libdwfl` does all the work on start and then provides a regular
`Dwarf*` accessor from `libdw`, which we already use. On start it
consumes about 257 ms for a Linux kernel binary compiled with DWARF5.

PiperOrigin-RevId: 495421080
Change-Id: Ie20542696e8e9e3e81a7f7ab152ac3be3407403f