Add fake symbols for types from DWARF
To test DWARF processing without implementing matching of DWARF types to
ELF symbols we need to attach STG nodes from DWARF entries to the root.
Hence, implement a temporary solution with attaching all nodes from
DWARF processors to the root `Symbols` node. This solution will be
replaced with actual matching of DWARF types to ELF symbols, which are
already attached to the root.

To generate "fake symbol" names, use a `Describe` functor, which creates
stable enough names for types. It is useful to reduce churn in test
outputs during addition of new types and migrating from temporary
solution to the permanent. But DWARF types, which are not used by the
ELF symbols, will be removed after such migration and we should
carefully review it to understand any loss of coverage.

PiperOrigin-RevId: 495429217
Change-Id: I7635f7c96800238ab0196b5cf7b92075e25fde5f