DWARF reading: process type definitions
Add support for DWARF entries tagged as `DW_TAG_typedef`, which is
referencing another DWARF entry without following this reference. Do
this by getting a `Dwarf_Die` (wrapped in `Entry`) from an attribute,
converting it to an offset and finally to an `Id` through
`GetIdForEntry`. This guarantees that all references would have the same
`Id` when being referenced and parsed.

There may exist some references that were not parsed since we haven't
implemented processing of all DWARF types. They are replaced with
`Variadic` node, but once the DWARF processor is stabilised, we will
replace it with `Die()` to shake out any bugs.

DWARF entries that do not have `DW_AT_type` attributes reference to the
`void` type. And DWARF does not have a dedicated entry for `void`, which
means we have to create an artificial `Void` node.

There is no other good way to get an offset from a reference directly,
because `dwarf_formref` was deprecated and doesn't support all forms
that `dwarf_formref_die` does.

Lastly, offset is only meaningful in the context of a single ELF
section. To support `.debug_types` or separate debug info files, we'll
need a different solution.

PiperOrigin-RevId: 495591888
Change-Id: I9eda8dd787f9318752b288e722214b278f45cda8