DWARF entry: add method to parse flag attributes
Add support for attributes of form `flag`, that can be "set" or
"absent". These states are mapped to `bool`: `true` "set" and `false`
for "absent".

In case of `DW_FORM_flag` it is possible to have the flag "absent" or
"set to false", but these are semantically identical in the DWARF
standard.

PiperOrigin-RevId: 495731880
Change-Id: Ifedecae2fe902d22601f033fc26b31de9d8614ea