proto: Fix `type_id` assignment for `ElfSymbol` in reader
Assign type IDs for `ElfSymbol` nodes using `Transformer::GetId()`
instead of `Transformer::Transform()`. Using the latter does not
allocate a new node in the graph and simply returns the value of the
external ID instead of an ID of a node in the graph.

Also, move the `protobuf/repeated_field.h` includes from
`proto_reader.h` to `proto_reader.cc`.

PiperOrigin-RevId: 495883694
Change-Id: I3afa1cbda6dd5f9a046ebcecc8dcc6fcc577b9e2