fingerprint: improve interface, support metrics
`Fingerprint` is now a standalone function that takes a graph and root
node id and returns a mapping of id to hash for all reachable nodes.

The hashes and todo set belong to this function and are passed to the
worker function object, now named `Hasher`.

The ad-hoc counter repointing has been replaced by `Metrics`.

PiperOrigin-RevId: 498146575
Change-Id: I599390354e3de79756c5443e26384ecd6c016a63