equality cache: featureful and instrumented equality cache
Recursive node equality is implemented by the `Equals` function object
but there is a lot of flexibility in how information about completed
comparisons can be used to speed up later ones. Optimal caching may
depend strongly on the application and intended inputs.

This equality cache is designed to support at least node deduplication
and has the following features, not all of which may be worth keeping
in the long term:

* hash comparison - distinct hash partitions contain unequal nodes
* union-find - to track sets of equal nodes
  * with path halving
  * and union by rank
* tracking inequalites between union-find roots

The cache is instrumented to record various metrics.

PiperOrigin-RevId: 498577579
Change-Id: I4b7f56b27a49a554a2372bcc371a8dd5c411aef3