deduplication: add graph deduplication facility
Inputs to STG often contain redundant information, resulting in some or
many duplicate nodes in the graph.

`Deduplicate` finds recursively-equal nodes in a graph. For each set
of equal nodes, it repoints references to all but one of these to the
one and removes the others.

The function takes a map whose keys are the ids of nodes to consider
and whose values are equality-respecting hashes. The hashes are used to
partition the nodes into sets of nodes that could be the same.

The function also takes and returns a root node id, as the original root
node could have been deduplicated.

PiperOrigin-RevId: 499514700
Change-Id: I822df3265fce3ad21713083208387f03e4ca4186