ELF loader: add support for CRC values by offset
New kernels changed the format of storing CRC values from absolute
symbol values to offsets in the `__kcrctab` or `__kcrctab_gpl` section.

Add a new method `GetElfSymbolCRC` to the ELF loader, which reads CRC
values in CRC sections by offset. It is very fast and doesn't affect the
performance of reading the ELF symbol table.

PiperOrigin-RevId: 499873253
Change-Id: Idf419369d26f51167feea4fa7c624a2d28f1239f