DWARF wrapper: check `elf_memory` returning `nullptr`
For fuzzing tests we introduced methods in DWFL to open ELF from memory.
But the new method `dwfl_report_offline_memory` doesn't check if the
`elf_memory` call completed successfully and tries to access data on
nullptr.

Add a workaround to check if `elf_memory` runs successfully before
passing data to the DWFL, but this should be replaced with a proper
error handling in `__libdw_open_elf_memory`.

PiperOrigin-RevId: 499903137
Change-Id: I7e8197e27c86b2fe2b5dfa17d82394074101aec6