DWARF reading: process members of structure
Add support for DWARF entries, tagged as `DW_TAG_member`. Members looks
mostly like typedefs, except the name is optional and the referred type
is required.

Members are linked to the parent structure using GetIdForEntry, despite
it being a direct child of this structure. It is more convenient to
treat the child-parent relation the same way as cross-type references.

Some `stg::Member` fields like `offset` and `bitsize` are not filled,
because their calculation is not trivial.

PiperOrigin-RevId: 500000137
Change-Id: I787d569e0075eb1da66462b13e2f2d086038c866