stg: add support for symbol filtering
Symbol filtering is restricting an ABI representation to some subset
of the exposed symbols.

This commit adds the `--symbols` (or `-S`) option which takes a symbol
filter expression. Symbols (including any `@version` or `@@version`
suffix) that match the expression are retained.

A basic expression is either a glob pattern or a `:` followed by the
name of a file containing a symbol list as accepted by libabigail. A
compound expression is formed using a boolean operator or parentheses
(`!`, `&`, `|`, `(` and `)`).

PiperOrigin-RevId: 500387172
Change-Id: I518585b19956e665b19c5696daffe7f6756c311f