DWARF reading: process variables
Add support for DWARF entries tagged as `DW_TAG_variable`. There is no
dedicated STG node for variables and this entry is used only for
matching DWARF type to the ELF symbol corresponding to the variable.

Variables (and functions in the future) are added to
`dwarf::Types::symbols`, which are used by the ELF reader. DWARF symbols
are not matched to STG symbols at this moment, so they are appended to
stg::Symbols using a "fake symbol name".

PiperOrigin-RevId: 501403019
Change-Id: I2ee4680f302cd92560f795e4aae5c73e8e2ec429