Remove struct / class distinction
Both `class` and `struct` types are now modelled as `struct`.

There are very few distinctions between `class foo` and `struct foo`,
particularly when it comes to ABI compatibility, as they are two names
for the same type. On the other hand, having multiple possible
representations and names for the same type complicates things like
forward declaration resolution and can result in unhelpful or even
misleading diffs if a change of tag (`struct` vs `class`) happens at the
same time as other changes.

We can consider reinstating the distinction, but it must be as a flag
that only influences descriptions, not semantics. This would be helpful
for users seeing `struct foo` in reports but not finding it in code that
only contains `class foo`.

PiperOrigin-RevId: 501831732
Change-Id: I66067fe2cd763418fa46e0938ee0d6a23874b060