ELF reading: read ELF symbols address
Add `ElfLoader::GetElfSymbolAddress` method to extract the address of
symbols. Most of the time `symbol.value` contains an absolute virtual
address, which is the same as in DWARF. But in relocatable files
`symbol.value` holds an offset in section of symbol. In this case we
need to get section start from section header and add it to the value of
the symbol to get the same absolute address, as in DWARF.

PiperOrigin-RevId: 502621718
Change-Id: I9e9e562b568f4008c0464526b1b7e44df974c742