add type resolution pass
This attempts to resolve each named (struct, union or enum) type's
definitions and declarations.

If there are one or more definitions for a named type and no conflicting
definitions, this will result in all definition and declaration nodes
being reduced to a single definition.

DWARF contains information about types, per compilation unit. C and C++
allow incomplete type definitions and different compilation units may
contain different mixes of named type declarations and definitions.
Given a large code base, there can be many compatible but different
definitions of a type across the DWARF compilation units.

These differences are resolved by recursively unifying pairs of
definitions. If a unification of definitions succeeds, each of the
recursively compared pairs of types (not just named types) are linked
with a mapping.

The only case where unification differs (locally) from equality is when
comparing a definition and declaration of the same named type.

The resolution pass is currently run just before (and in the case of
DWARF inputs, does much of the work of) the deduplication pass. This may
change as it fairly specific to DWARF handling.

PiperOrigin-RevId: 503178172
Change-Id: I9b709daa8d7610ae4c8d071bec2a6ecc311947eb