DWARF reading: support a bit field offset and size
Add support for a bit field offset adjustment. Before the change the
offset was the number of bits before the bit field container. Now the
code accounts for the offset inside the bit field container.

Pass an endianness of the binary to the DWARF processor, because the
adjustment depends on it.

Calculating the bit field offset uses a bit field size, which is also
part of the ABI. So it is reasonable to pass it in an STG node.

PiperOrigin-RevId: 504029959
Change-Id: I79acab4eccf393d5a420d37e50a93064e46704a2