graph: multiple vector representation
The virtual base class is removed. There is now a separate vector of
nodes for each kind of node and an indirection vector containing kind,
offset pairs. Function dispatch now uses switch statements instead of
RTTI.

STG's `Graph` and node types and the `Apply` functions together provide
an abstraction to the rest of the code base. The concrete representation
can be varied, so long as the interface remains the same.

We experimented with multiple concrete representations (including
combinations of `std::variant` / `std::visit`, pointers and values).
Some were more memory efficient. Some made graph node addition faster,
at the expense of slower graph processing.

We also looked at replacing `std::vector` with `std::unordered_map` to
better support node deallocation, but this consumed significantly more
peak memory and was slower.

The concrete representation here is Pareto optimal and believed to be
the fastest of the main variations. It's around 5% faster than the
status quo.

Further possibilities include:

* halving the number of bits in a node ID (making lots of things
  smaller)
* packing node kind and offset together (shrinking the indirection
  vector)

We could do both, but we'd only be able to support 2^28 nodes of each
kind and 2^32 nodes altogether.

PiperOrigin-RevId: 504257348
Change-Id: I92de8e02094074f7c66ef9d0baee44a98968c7ae