model enum underlying types
The DWARF standard allows `DW_TAG_enumeration_type` information entities
to have an underlying `DW_AT_type`. GCC and Clang always emit this and
it is needed for C++ where the underlying type can be fixed as in `enum
X : char`.

Separately, DWARF allows the size of an enumeration type to be specified
as an attribute of the enum itself. Given the underlying type, this is
always redundant.

Until now, STG has not modelled enum-underlying types. This is partly
because BTF doesn't and libabigail does so incompletely (the names of
such types are lost). BTF does provide sign and size information as part
of the enum itself, while libabigail provides underlying primitive types
(which always lack encoding) with a synthetic name and correct size.

With this change:

* each enum definition includes an underlying type
* the BTF reader synthesises an underlying type with suitable name, size
  and encoding
* the Abigail reader builds the given underlying type
* the native reader and writer expect and emit underlying types
* `Fingerprint` finds and fingerprints underlying types
* `Substitute` updates underlying types
* `Compare`, `Equals` and `Unify` all compare underlying types

PiperOrigin-RevId: 504278787
Change-Id: I45b46e84ce4709b5a0d0a45abcdf78b1c9a16c36