stgdiff: remove --compare-options all support
`stgdiff` must be invoked with an explicit list of compare options, if
needed.

`stgdiff` can compare XML, BTF and ELF/DWARF inputs. The ABI graphs we
get from each source are not identical. Compare options are a way to
reduce comparison noise. The `all` option was a useful default for
XML-XML comparisons when used in conjunction with other monitoring to
detect loss of type information. It is not a useful default for other
combinations of inputs.

PiperOrigin-RevId: 504572703
Change-Id: Ic0bb9cc816aa04837c4107fbf1aa8ffacb2df390