ELF reader: consider name when matching ELF to DWARF
Some symbols can have the same address, but relate to different objects. For
example, if one of the symbols has zero size another symbol can start at the
same address.

Such collisions can be resolved by symbol name, because it is impossible to have
multiple symbols with the same name. But if we do find different DWARF types
for the same name and address we should investigate, so fail.

PiperOrigin-RevId: 506926957
Change-Id: I64251a552143bb617d939cae7c1f175512ae4f47