comparison: replace CompareOptions with Ignore
The internal renaming is because CompareOptions only contained two options to
ignore certain differences and all comparison options should be of this form.

The new representation is an enum-encoded bitset which permits easy addition
(and removal) of options and is small enough to pass around by value.

This change also makes `CompareDefined` a method of `Compare`, changing how it
gets the option information.

PiperOrigin-RevId: 508315106
Change-Id: Idc192fef87fc1d0592ed844d1b7844eb40323c6b