type resolution: use specialised data structures to track seen nodes and mappings
Type resolution is currently the most expensive processing phase of STG, after
DWARF extraction. A significant fraction of the cost is associated with
inserting and looking up node IDs in various sets and maps and two of the tables
grow to essentially include all IDs.

Change their types as follows:

* `NamedTypes::seen`: `unordered_set<Id>` becomes `DenseIdSet`
   * this is a wrapped `vector<bool>`
* `UnificationCache::mapping`: `unordered_map<Id, Id>` becomes `DenseIdMapping`
   * this is a wrapped `vector<Id>`

The structures are fully populated on construction. Any out-of-bounds
access will result in an expensive resize operation.

Extracting STG from a Linux kernel is about 25% faster with this change.

PiperOrigin-RevId: 512600919
Change-Id: I2cf6f5df347954765ecb08bf6039a795cde412c0